% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{mixsiar_save}
\alias{mixsiar_save}
\title{Save MixSIAR output in new directory}
\usage{
mixsiar_save(
  new_dir = "mixing_model_output",
  isospace_filename = "isospace",
  model_filename = "mixsiar_model.txt"
)
}
\arguments{
\item{new_dir}{name of directory to store output from a MixSIAR model}

\item{isospace_filename}{Name of the isospace plot without the "_1_2" which
MixSIAR appends}

\item{model_filename}{Name of the model filename}
}
\value{
A subdirectory with the output from a MixSIAR model
}
\description{
This function makes it easier to organize output from multiple MixSIAR models
by saving the output in a named subdirectory of output's original directory
}
\examples{
mixsiar_save(new_dir = "main_model",
             isospace_filename = "isospace",
             model_filename = "mixsiar_model.txt")
}
